Harvest executor refactor
