+++ src/main/scala/dpla/ingestion3/entries/ingest/HarvestEntry.scala
@@ -1,14 +1,8 @@
 package dpla.ingestion3.entries.ingest
 
 import dpla.ingestion3.confs.{CmdArgs, Ingestion3Conf, i3Conf}
-import dpla.ingestion3.harvesters.Harvester
-import dpla.ingestion3.harvesters.oai.OaiHarvester
-import dpla.ingestion3.harvesters.pss.PssHarvester
-import dpla.ingestion3.harvesters.resourceSync.RsHarvester
-import dpla.ingestion3.utils.{ProviderRegistry, Utils}
-import org.apache.log4j.Logger
-
-import scala.util.{Failure, Success}
+import dpla.ingestion3.executors.HarvestExecutor
+import dpla.ingestion3.utils.Utils
 
 /**
   * Entry point for running a harvest.
@@ -18,7 +12,7 @@ import scala.util.{Failure, Success}
   *   --conf    Path to conf file
   *   --name    Provider short name
   */
-object HarvestEntry {
+object HarvestEntry extends HarvestExecutor {
 
   def main(args: Array[String]): Unit = {
 
@@ -37,66 +31,6 @@ object HarvestEntry {
     val providerConf: i3Conf = i3Conf.load()
 
     // Execute harvest.
-    executeHarvest(shortName, dataOut, providerConf, harvestLogger)
-  }
-
-  /**
-    * Run the appropriate type of harvest.
-    *
-    * @param shortName Provider short name (e.g. cdl, mdl, nara, loc).
-    *                  @see ProviderRegistry.register() for the authoritative
-    *                       list of provider short names.
-    * @param outputDir Location to save output from harvest
-    * @param conf Configurations read from application configuration file
-    * @param harvestLogger Logger object
-    */
-  def executeHarvest(shortName: String,
-                     outputDir: String,
-                     conf: i3Conf,
-                     harvestLogger: Logger) = {
-
-    // Log config file location and provider short name.
-    harvestLogger.info(s"Harvest initiated")
-    harvestLogger.info(s"Provider short name: $shortName")
-
-    // Get and log harvest type.
-    val harvestType = conf.harvest.harvestType
-      .getOrElse(throw new RuntimeException("No harvest type specified."))
-
-    harvestLogger.info(s"Harvest type: $harvestType")
-
-    val harvester: Harvester = harvestType match {
-      case "oai" =>
-        new OaiHarvester(shortName, conf, outputDir, harvestLogger)
-      case "api" =>
-        registeredHarvester(shortName, outputDir, conf, harvestLogger)
-      case "pss" =>
-        new PssHarvester(shortName, conf, outputDir, harvestLogger)
-      case "rs" =>
-        new RsHarvester(shortName, conf, outputDir, harvestLogger)
-      case _ =>
-        val msg = s"Harvest type not recognized."
-        harvestLogger.fatal(msg)
-        throw new RuntimeException(msg)
-    }
-
-    harvester.harvest
-  }
-
-  // Look up a registered Harvester class with the given shortName and instantiate.
-  def registeredHarvester(shortName: String,
-                          outputDir: String,
-                          conf: i3Conf,
-                          harvestLogger: Logger): Harvester = {
-
-    val harvesterClass = ProviderRegistry.lookupHarvesterClass(shortName) match {
-      case Success(harvClass) => harvClass
-      case Failure(e) =>
-        harvestLogger.fatal(e.getMessage)
-        throw e
-    }
-
-    harvesterClass.getConstructor(classOf[String], classOf[i3Conf], classOf[String], classOf[Logger])
-      .newInstance(shortName, conf, outputDir, harvestLogger)
+    execute(shortName, dataOut, providerConf, harvestLogger)
   }
 }
